add discord webhook (#13)
