+++ src/main/java/org/starcoin/poll/api/service/WebhookService.java
@@ -1,0 +1,5 @@
+package org.starcoin.poll.api.service;
+
+public interface WebhookService {
+    void post(String subject, String text);
+}
